% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{freq_value}
\alias{freq_value}
\title{freq_value}
\usage{
freq_value(x)
}
\description{
This function determines the most frequently occurring value in an integer
vector (mode). If the mode is ambiguous, a function returns any mode.
}
\details{
ARGUMENTS:
x - a integer vector

RETURN VALUE:
a integer value
}
\seealso{
Other fungroup: \code{\link{list2array}};
  \code{\link{perms}}
}

