% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perms}
\alias{perms}
\title{perms}
\usage{
perms(x)
}
\description{
This function generates all the possible permutations of the set
{1, 2, ..., n} (for some given n). A function removes missing values.
}
\details{
ARGUMENTS:
x - a integer vector

RETURN VALUE:
a matrix with n! rows and n columns
}
\seealso{
Other fungroup: \code{\link{freq_value}};
  \code{\link{list2array}}
}

